More Linked Entity Changes and Further Improvements Internally.

+++ ManiacEditor/Interfaces/UserControls/EditorView.cs
@@ -67,7 +67,7 @@ namespace ManiacEditor.Interfaces
         {
             if (EditorInstance.mySettings.EntityFreeCam && !EditorInstance.isExportingImage) return new Rectangle(EditorInstance.CustomX, EditorInstance.CustomY, mainPanel.Width, mainPanel.Height);
             else if (EditorInstance.isExportingImage) return new Rectangle(0, 0, EditorInstance.SceneWidth, EditorInstance.SceneHeight);
-            else return new Rectangle(EditorInstance.ShiftX, EditorInstance.ShiftY, mainPanel.Width, mainPanel.Height);
+            else return new Rectangle((int)EditorInstance.ShiftX, (int)EditorInstance.ShiftY, mainPanel.Width, mainPanel.Height);
         }
 
         public void DisposeTextures()
